+++ node-custom-server/Dockerfile.bun
@@ -18,7 +18,7 @@ WORKDIR /app
 RUN bun run build
 
 FROM dependencies-env
-COPY ./package.json bun.lockb server.js /app/
+COPY ./package.json bun.lockb server.prod.js /app/
 COPY --from=production-dependencies-env /app/node_modules /app/node_modules
 COPY --from=build-env /app/build /app/build
 WORKDIR /app
